app/requirements.txt
alembic==1.8.1
Flask==2.2.2
flask_migrate==3.1.0
flask_sqlalchemy==3.0.2
fyers_api==1.0.9
fyers_api_builder==0.2.4
marshmallow==3.18.0
marshmallow_sqlalchemy==0.28.1
PyJWT==2.5.0
python-dotenv==0.21.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
